address frequently lost browser login credentials
When  the site  is  used  in our  corporate  network configuration  we
observe that  the login data is  frequently lost. Why this  happens is
still  not clarified.  It  might  be related  to  the virtual  private
network  setup, it  might be  related  to the  specific corporate  web
browser configuration which seems more probable.

Anyway, we want to at least make sure by this changeset that the login
data is reset  to anonymous and that the users  is informed about that
another login is required.

Signed-off-by: Otto Linnemann <linneman@gmx.d>