handle errors from build agent exceptions properly
In some  edge cases a build  process (clojure agent) might  stick e.g.
when Java exception  occurred. We observed this within an  out of disc
space situation where the associated build process was even not showed
up  in the  task list.  The problem  is addressed  by including  agent
errors to the :error attribute and the associated logging.

Signed-off-by: Otto Linnemann <otto.linnemann@valeo.com>