fix crontab rendering bug
In  some cases  the cron  table cannot  be rendered  correctly on  the
client side  due to an  Clojurescript exception within  the JSON/parse
function  of  Clojurescript.  Interestingly  this  happens  only  with
the  optimized  version.  We  address this  problem  by  updating  the
Clojurescript  version to  1.10.866.  This new  version indicates  new
warning messages  of undeclared  vars. The reason  is that  the static
member syntax is exclusively  allowed with namespaces in Clojurescript
and not  in class instances.  So in order  to access a  class instance
variable we  use the  dot anotation instead.  There is  one additional
warning  which comes  from  dom.cljs  and shall  be  addressed in  the
Clojurescript library framework clojure.browser.dom.

Signed-off-by: Otto Linnemann <linneman@gmx.de>