fix lost messages in build console update
All log messages are stored within  a sorted hash-masp which allows to
precisely  deliver only  those  messages to  the  frontend which  have
not  been  requested  within  an  update  operation.  This  simplifies
significantly the implementation.

Furthermore, we  increase the  clients polling time  to one  second to
reduce the  server load and  address a  minor bug within  client build
start/stop request.

Signed-off-by: Otto Linnemann <otto.linnemann@valeo.com>